labels type long

+++ simple_nn_runner.py
@@ -78,8 +78,8 @@ test_features = torch.tensor(X_test.values.astype(np.float32))
 
 # print(train_features.shape)
 # print(le.transform(y_train).shape)
-train_dataset = TensorDataset(train_features, torch.tensor(le.transform(y_train).astype(np.float32)))
-test_dataset = TensorDataset(test_features, torch.tensor(le.transform(y_test).astype(np.float32)))
+train_dataset = TensorDataset(train_features, torch.tensor(le.transform(y_train)))
+test_dataset = TensorDataset(test_features, torch.tensor(le.transform(y_test)))
 
 train_loader = DataLoader(train_dataset, batch_size=64,
 #  collate_fn=collate_wrapper,
